+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ##### Breaking
 
-- None.
+- The `--xcargs` option has been removed, and superseded by passing arguments following the `--` terminator. E.g `periphery scan --xcargs --foo` is now `periphery scan -- --foo`. This feature can also be used to pass arguments to `swift build` for SwiftPM projects.
 
 ##### Enhancements